refactor: RestaurantCard 분리
