+++ src/index.js
@@ -3,8 +3,10 @@ import WebController from './Controller/WebController';
 
 import Dropdown from './view/components/Dropdown';
 import RestaurantCards from './view/components/RestaurantCards';
+import RestaurantCard from './view/components/RestaurantCard';
 
 window.customElements.define('drop-down', Dropdown, { extends: 'select' });
 window.customElements.define('restaurant-cards', RestaurantCards, { extends: 'ul' });
+window.customElements.define('restaurant-card', RestaurantCard, { extends: 'li' });
 
 new WebController().run();
