+++ src/view/components/RestaurantCards.js
@@ -1,13 +1,5 @@
 import restaurantCatalog, { SORT_CONDITION } from '../../domain/RestaurantCatalog';
-
-const IMG_CATEGORY = Object.freeze({
-  한식: 'korean',
-  아시안: 'asian',
-  중식: 'chinese',
-  기타: 'etc',
-  양식: 'western',
-  일식: 'japanese',
-});
+import RestaurantCard from './RestaurantCard';
 
 const [SORT_BY_NAME, SORT_BY_DISTANCE] = SORT_CONDITION;
 
@@ -17,6 +9,8 @@ function sortMethod(attribute) {
 }
 
 class RestaurantCards extends HTMLUListElement {
+  #renderedRestaurants = [];
+
   #appendRestaurants() {
     const dataFilteredAndSorted = sortMethod(this.getAttribute('data-sort-select'))(
       restaurantCatalog
@@ -28,24 +22,9 @@ class RestaurantCards extends HTMLUListElement {
 
   #appendRestaurantElement(restaurants) {
     restaurants.forEach((data) => {
-      const liElement = document.createElement('li');
-      liElement.classList.add('restaurant');
-      liElement.innerHTML = this.#generateInnerHTML(data);
+      const liElement = new RestaurantCard(data);
       this.appendChild(liElement);
     });
-  }
-
-  #generateInnerHTML({ category, name, distanceFromCampus, description }) {
-    return `
-    <div class="restaurant__category">
-      <img src="./templates/category-${IMG_CATEGORY[category]}.png" alt="${category}" class="category-icon">
-    </div>
-    <div class="restaurant__info">
-      <h3 class="restaurant__name text-subtitle">${name}</h3>
-      <span class="restaurant__distance text-body">캠퍼스부터 ${distanceFromCampus}분 내</span>
-      <p class="restaurant__description text-body">${description}</p>
-    </div>
-  `;
   }
 
   clear() {
